generalize queue manager for re-using with other message queues
