+++ src/test/java/com/github/libgraviton/workerbase/mq/WorkerQueueManagerTest.java
@@ -1,0 +1,56 @@
+package com.github.libgraviton.workerbase.mq;
+
+import com.github.libgraviton.workerbase.helper.PropertiesLoader;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.util.Properties;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.times;
+
+public class WorkerQueueManagerTest {
+
+    private WorkerQueueManager queueManager;
+
+    private Properties properties;
+
+    @Before
+    public void setup() throws Exception {
+        properties = PropertiesLoader.load();
+    }
+
+    @Test
+    public void shouldInitializeQueueManager() {
+        queueManager = new WorkerQueueManager(properties);
+        assertEquals("document.core.app.*", queueManager.getBindKeys().get(0));
+        assertEquals("graviton", queueManager.getExchangeName());
+        assertEquals("aaaaaaaaaaaaaa", queueManager.getFactory().getHost());
+        assertEquals(5672, queueManager.getFactory().getPort());
+        assertEquals(1, queueManager.getRetryAfterSeconds());
+    }
+
+    @Test
+    public void shouldTryConnectingToQueue() {
+        queueManager = spy(new WorkerQueueManager(properties));
+        WorkerQueueConnector queueConnector = spy(new WorkerQueueConnector());
+        queueConnector.setBindKeys(queueManager.getBindKeys());
+        queueConnector.setRetryAfterSeconds(queueManager.getRetryAfterSeconds());
+        queueConnector.setWorker(queueManager.getWorker());
+        queueConnector.setFactory(queueManager.getFactory());
+        queueConnector.setExchangeName(queueManager.getExchangeName());
+
+        doCallRealMethod().doReturn(true).when(queueConnector).connectAttempt();
+        doReturn(queueConnector).when(queueManager).getQueueConnector();
+
+        verify(queueConnector, times(0)).connectAttempt();
+        queueManager.connect();
+        try {
+            Thread.sleep(1500);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        verify(queueConnector, times(2)).connectAttempt();
+    }
+}
